third_party/WebKit/LayoutTests/inspector/elements/styles/empty-background-url.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<link rel="stylesheet" href="resources/empty-background-url.css">
<script>
function test()
{
    InspectorTest.selectNodeAndWaitForStylesWithComputed("inspected", step1);

    function step1()
    {
        InspectorTest.dumpSelectedElementStyles(true, false);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that empty url in the property value does not break inspector.</p>
<div id="inspected"></div>
</body>
</html>
