third_party/WebKit/LayoutTests/inspector/elements/styles/empty-background-url-expected.txt

Tests that empty url in the property value does not break inspector.

[expanded] 
element.style  { ()

[expanded] 
#inspected  { (empty-background-url.css:1 -> empty-background-url.css:1:1)
background-image: url();

[expanded] 
div  { (user agent stylesheet)
display: block;
